Implemented 30second-games
http://en.lichess.org/forum/lichess-feedback/request-new-time-control-30-seconds0
http://en.lichess.org/forum/team-30-seconds--0-time-control/this-is-how-we-can-get-0-games

+++ modules/setup/src/main/AiConfig.scala
@@ -7,7 +7,7 @@ import lila.lobby.Color
 case class AiConfig(
     variant: chess.variant.Variant,
     timeMode: TimeMode,
-    time: Int,
+    time: Double,
     increment: Int,
     days: Int,
     level: Int,
@@ -40,7 +40,7 @@ case class AiConfig(
 
 object AiConfig extends BaseConfig {
 
-  def <<(v: Int, tm: Int, t: Int, i: Int, d: Int, level: Int, c: String, fen: Option[String]) = new AiConfig(
+  def <<(v: Int, tm: Int, t: Double, i: Int, d: Int, level: Int, c: String, fen: Option[String]) = new AiConfig(
     variant = chess.variant.Variant(v) err "Invalid game variant " + v,
     timeMode = TimeMode(tm) err s"Invalid time mode $tm",
     time = t,
@@ -53,7 +53,7 @@ object AiConfig extends BaseConfig {
   val default = AiConfig(
     variant = variantDefault,
     timeMode = TimeMode.Unlimited,
-    time = 5,
+    time = 5d,
     increment = 8,
     days = 2,
     level = 1,
@@ -71,7 +71,7 @@ object AiConfig extends BaseConfig {
     def reads(r: BSON.Reader): AiConfig = AiConfig(
       variant = chess.variant.Variant orDefault (r int "v"),
       timeMode = TimeMode orDefault (r int "tm"),
-      time = r int "t",
+      time = r double "t",
       increment = r int "i",
       days = r int "d",
       level = r int "l",
